+++ src/modules/OpponentRankExtractor.cpp
@@ -9,7 +9,7 @@ int matchParenStart(const cv::Mat &lineBinary, const cv::Mat &parenStart)
 		cv::Rect matchingRect(i, 0, parenStart.cols, parenStart.rows);
 		cv::Mat matchResult = parenStart ^ lineBinary(matchingRect);
 		const double difference = cv::sum(matchResult)[0] / 255.0;
-		if (difference < parenStart.total() * 0.05) {
+		if (difference < static_cast<double>(parenStart.total()) * 0.05) {
 			return i + parenStart.cols;
 		}
 	}
@@ -23,7 +23,7 @@ int matchParenEnd(const cv::Mat &lineBinary, const cv::Mat &parenEnd,
 		cv::Rect matchingRect(i, 0, parenEnd.cols, parenEnd.rows);
 		cv::Mat matchResult = parenEnd ^ lineBinary(matchingRect);
 		const double difference = cv::sum(matchResult)[0] / 255.0;
-		if (difference < parenEnd.total() * 0.05) {
+		if (difference < static_cast<double>(parenEnd.total()) * 0.05) {
 			return i;
 		}
 	}
